Added mongodb support
