+++ src/main/java/com/surveyMonkey/entities/OptionQuestion.java
@@ -1,5 +1,7 @@
 package com.surveyMonkey.entities;
 
+import org.springframework.data.mongodb.core.mapping.Document;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -9,7 +11,7 @@ import javax.persistence.Entity;
 
 import static com.surveyMonkey.util.Constants.OPTION;
 
-@Entity
+@Document
 public class OptionQuestion extends Question {
 
     private static final long serialVersionUID = 1795113345322280778L;
