+++ src/main/java/com/surveyMonkey/entities/Survey.java
@@ -1,5 +1,7 @@
 package com.surveyMonkey.entities;
 
+import org.springframework.data.mongodb.core.mapping.Document;
+
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
@@ -12,19 +14,18 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
-@Entity
+@Document
 public class Survey implements Serializable {
 
     private static final long serialVersionUID = -7426934374543805936L;
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private String id;
     @OneToMany(cascade = CascadeType.ALL)
     private List<QuestionAnswerWrapper> survey = new ArrayList<>();
     private String title;
     private String surveyCode;
     private String surveyPassword;
-    private final int SURVEY_CODE_LENGTH = 5;
+    private int SURVEY_CODE_LENGTH = 5;
 
     public Survey() {
     }
@@ -39,11 +40,11 @@ public class Survey implements Serializable {
         survey.add(new QuestionAnswerWrapper(q));
     }
 
-    public Long getId() {
+    public String getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(String id) {
         this.id = id;
     }
