+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+#Local MongoDB config
+spring.data.mongodb.authentication-database=admin
+spring.data.mongodb.database=test
+spring.data.mongodb.port=27017
+spring.data.mongodb.host=localhost
